chart - move config to parent
