+++ src/views/StatsOverviewMemos.vue
@@ -7,7 +7,7 @@
       <div class="bg-white flex items-center justify-center min-h-full relative flex-shrink-0 py-6">
         <div class="w-full">
           <div class="mx-auto px-12">
-            <BaseChart :options="{}" />
+            <BaseChart :config="getChartConfig({ label: 'Users', data: usersData })" />
             <br />
             <br />
             <br />
@@ -30,8 +30,84 @@ import NavigationBottom from '@/components/BaseNavigationBottom';
 import BaseAppBarHeader from '@/components/BaseAppBarHeader.vue';
 import BaseChart from '@/components/BaseChart.vue';
 
+import { colors } from '@/constants/theme.js';
+
 export default {
   name: 'StatsOverviewMemos',
-  components: { BaseAppBarHeader, NavigationBottom, LayoutFixedFooter, BaseChart }
+  components: { BaseAppBarHeader, NavigationBottom, LayoutFixedFooter, BaseChart },
+  data: () => ({
+    // TODO: read from store
+    usersData: [100, 500, 100, 200, 300, 800, 900]
+  }),
+  methods: {
+    getChartConfig({ label, data }) {
+      const config = {
+        type: 'line',
+        data: {
+          labels: ['M', 'T', 'W', 'Th', 'F', 'S', 'Su'],
+          datasets: [
+            {
+              label,
+              data,
+              borderWidth: 2,
+              backgroundColor: colors.secondary,
+              borderColor: colors.secondary,
+              fill: false
+            }
+          ]
+        },
+        options: {
+          elements: {
+            point: {
+              radius: 0
+            }
+          },
+          legend: {
+            display: false
+          },
+          tooltips: false,
+          responsive: true,
+          hover: {
+            mode: 'nearest',
+            intersect: true
+          },
+          scales: {
+            yAxes: [
+              {
+                ticks: { min: 0, display: false },
+                gridLines: {
+                  drawBorder: false,
+                  display: false
+                }
+              }
+            ],
+            xAxes: [
+              {
+                position: 'bottom',
+                gridLines: {
+                  drawBorder: false,
+                  lineWidth: 1,
+                  color: colors.divider
+                }
+              },
+              {
+                position: 'top',
+                gridLines: {
+                  drawBorder: false,
+                  display: false
+                },
+                ticks: {
+                  callback: function(value, index, values) {
+                    return data[index];
+                  }
+                }
+              }
+            ]
+          }
+        }
+      };
+      return config;
+    }
+  }
 };
 </script>
